Components!
Pretty large update, which adds some new functionalities.
- Entities are now collections of various components that one can
assemble together.
- Input no longer directly triggers entity commands- they pass input to
an input handler given, which pass input to their object.
Press states- a regular press (with repeat detection) a hold (which just
checks if the key is pressed) and a release (with repeat detection, but
is not too necessary)
- Converted all .h files to equivalent .hpp files. Nothing too
necessary, but it helps to differentiate them as C++ headers and not C
headers.
- DId a bunch of reformatting, namely getting rid of all the spaced-out
tabs and replacing tem with actual tabs
- Added a third sprite because why the hell not